# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Pragati Path API","description":"Backend API for the Student Academic Risk Early Warning System","version":"2.0.0"},"paths":{"/api/v1/datasets/upload":{"post":{"summary":"Upload Dataset","operationId":"upload_dataset_api_v1_datasets_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_dataset_api_v1_datasets_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DatasetResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/datasets/{dataset_id}/inspect":{"get":{"summary":"Inspect Dataset","operationId":"inspect_dataset_api_v1_datasets__dataset_id__inspect_get","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"integer","title":"Dataset Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/datasets/{dataset_id}/clean":{"post":{"summary":"Clean Dataset","operationId":"clean_dataset_api_v1_datasets__dataset_id__clean_post","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"integer","title":"Dataset Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CleaningReport"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/datasets/{dataset_id}/eda":{"post":{"summary":"Generate Eda","operationId":"generate_eda_api_v1_datasets__dataset_id__eda_post","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"integer","title":"Dataset Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/datasets/{dataset_id}/train":{"post":{"summary":"Train Models","operationId":"train_models_api_v1_datasets__dataset_id__train_post","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"integer","title":"Dataset Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/models/{model_id}/comparison":{"get":{"summary":"Get Model Comparison","operationId":"get_model_comparison_api_v1_models__model_id__comparison_get","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"integer","title":"Model Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/datasets/{dataset_id}/predict":{"post":{"summary":"Predict","operationId":"predict_api_v1_datasets__dataset_id__predict_post","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"integer","title":"Dataset Id"}},{"name":"model_id","in":"query","required":true,"schema":{"type":"integer","title":"Model Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StudentRiskResponse"},"title":"Response Predict Api V1 Datasets  Dataset Id  Predict Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/datasets/{dataset_id}/students":{"get":{"summary":"Get Students","operationId":"get_students_api_v1_datasets__dataset_id__students_get","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"integer","title":"Dataset Id"}},{"name":"search","in":"query","required":false,"schema":{"type":"string","title":"Search"}},{"name":"class_name","in":"query","required":false,"schema":{"type":"string","title":"Class Name"}},{"name":"risk_level","in":"query","required":false,"schema":{"type":"string","title":"Risk Level"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StudentRiskResponse"},"title":"Response Get Students Api V1 Datasets  Dataset Id  Students Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/datasets/{dataset_id}/students/{student_id}":{"get":{"summary":"Get Student","operationId":"get_student_api_v1_datasets__dataset_id__students__student_id__get","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"integer","title":"Dataset Id"}},{"name":"student_id","in":"path","required":true,"schema":{"type":"integer","title":"Student Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentRiskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/datasets/{dataset_id}/students/{student_id}/review":{"patch":{"summary":"Review Student","operationId":"review_student_api_v1_datasets__dataset_id__students__student_id__review_patch","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"integer","title":"Dataset Id"}},{"name":"student_id","in":"path","required":true,"schema":{"type":"integer","title":"Student Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FacultyReviewUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentRiskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/models/{model_id}/what-if":{"post":{"summary":"What If Prediction","operationId":"what_if_prediction_api_v1_models__model_id__what_if_post","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"integer","title":"Model Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WhatIfRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/datasets/{dataset_id}/analytics/class":{"get":{"summary":"Get Class Analytics","operationId":"get_class_analytics_api_v1_datasets__dataset_id__analytics_class_get","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"integer","title":"Dataset Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClassAnalyticsResponse"},"title":"Response Get Class Analytics Api V1 Datasets  Dataset Id  Analytics Class Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/datasets/{dataset_id}/export":{"get":{"summary":"Export Students Csv","operationId":"export_students_csv_api_v1_datasets__dataset_id__export_get","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"integer","title":"Dataset Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/datasets/{dataset_id}/dashboard-summary":{"get":{"summary":"Dashboard Summary","operationId":"dashboard_summary_api_v1_datasets__dataset_id__dashboard_summary_get","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"integer","title":"Dataset Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DashboardSummary"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_dataset_api_v1_datasets_upload_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_dataset_api_v1_datasets_upload_post"},"ClassAnalyticsResponse":{"properties":{"class_name":{"type":"string","title":"Class Name"},"student_count":{"type":"integer","title":"Student Count"},"average_attendance":{"type":"number","title":"Average Attendance"},"average_score":{"type":"number","title":"Average Score"},"high_risk_count":{"type":"integer","title":"High Risk Count"}},"type":"object","required":["class_name","student_count","average_attendance","average_score","high_risk_count"],"title":"ClassAnalyticsResponse"},"CleaningReport":{"properties":{"dataset_id":{"type":"integer","title":"Dataset Id"},"rows_before":{"type":"integer","title":"Rows Before"},"rows_after":{"type":"integer","title":"Rows After"},"duplicates_removed":{"type":"integer","title":"Duplicates Removed"},"missing_values_handled":{"type":"integer","title":"Missing Values Handled"},"columns_transformed":{"items":{"type":"string"},"type":"array","title":"Columns Transformed"}},"type":"object","required":["dataset_id","rows_before","rows_after","duplicates_removed","missing_values_handled","columns_transformed"],"title":"CleaningReport"},"DashboardSummary":{"properties":{"total_students":{"type":"integer","title":"Total Students"},"high_risk_count":{"type":"integer","title":"High Risk Count"},"medium_risk_count":{"type":"integer","title":"Medium Risk Count"},"low_risk_count":{"type":"integer","title":"Low Risk Count"},"average_attendance":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Average Attendance"},"average_performance":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Average Performance"}},"type":"object","required":["total_students","high_risk_count","medium_risk_count","low_risk_count","average_attendance","average_performance"],"title":"DashboardSummary"},"DatasetResponse":{"properties":{"id":{"type":"integer","title":"Id"},"filename":{"type":"string","title":"Filename"},"status":{"type":"string","title":"Status"},"upload_time":{"type":"string","format":"date-time","title":"Upload Time"}},"type":"object","required":["id","filename","status","upload_time"],"title":"DatasetResponse"},"FacultyReviewUpdate":{"properties":{"reviewed":{"type":"boolean","title":"Reviewed"},"review_notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Review Notes"},"intervention_action":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Intervention Action"}},"type":"object","required":["reviewed"],"title":"FacultyReviewUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InspectionResponse":{"properties":{"dataset_id":{"type":"integer","title":"Dataset Id"},"columns_found":{"items":{"type":"string"},"type":"array","title":"Columns Found"},"mapped_columns":{"additionalProperties":{"type":"string"},"type":"object","title":"Mapped Columns"},"unmapped_required":{"items":{"type":"string"},"type":"array","title":"Unmapped Required"},"can_proceed":{"type":"boolean","title":"Can Proceed"}},"type":"object","required":["dataset_id","columns_found","mapped_columns","unmapped_required","can_proceed"],"title":"InspectionResponse"},"ModelMetrics":{"properties":{"model_type":{"type":"string","title":"Model Type"},"accuracy":{"type":"number","title":"Accuracy"},"precision":{"type":"number","title":"Precision"},"recall":{"type":"number","title":"Recall"},"f1_score":{"type":"number","title":"F1 Score"}},"type":"object","required":["model_type","accuracy","precision","recall","f1_score"],"title":"ModelMetrics"},"StudentRiskResponse":{"properties":{"id":{"type":"integer","title":"Id"},"student_identifier":{"type":"string","title":"Student Identifier"},"class_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Class Name"},"attendance":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Attendance"},"latest_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Latest Score"},"risk_level":{"type":"string","title":"Risk Level"},"probability":{"type":"number","title":"Probability"},"explanation":{"type":"string","title":"Explanation"},"reviewed":{"type":"boolean","title":"Reviewed"},"review_notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Review Notes"},"intervention_action":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Intervention Action"},"original_data":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Original Data"}},"type":"object","required":["id","student_identifier","class_name","attendance","latest_score","risk_level","probability","explanation","reviewed","review_notes","intervention_action","original_data"],"title":"StudentRiskResponse"},"TrainResponse":{"properties":{"dataset_id":{"type":"integer","title":"Dataset Id"},"models_evaluated":{"items":{"$ref":"#/components/schemas/ModelMetrics"},"type":"array","title":"Models Evaluated"},"best_model":{"$ref":"#/components/schemas/ModelMetrics"},"model_id":{"type":"integer","title":"Model Id"}},"type":"object","required":["dataset_id","models_evaluated","best_model","model_id"],"title":"TrainResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WhatIfRequest":{"properties":{"original_data":{"type":"string","title":"Original Data"},"attendance":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Attendance"},"latest_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Latest Score"}},"type":"object","required":["original_data"],"title":"WhatIfRequest"}}}}